team-astra-platform/backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.6
fastapi==0.135.3
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.3.2
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
idna==3.11
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
pypdfium2==5.6.0
python-dotenv==1.2.2
python-multipart==0.0.22
requests==2.33.1
SQLAlchemy==2.0.49
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.43.0
